src: organize into module directories
