+++ src/models/chars.ts
