+++ src/services/index.ts
@@ -1,0 +1,3 @@
+export { Config, getConfig } from "./config";
+export { Git } from "./git";
+export { Logger } from "./logger";
